fleetpulse-dispatcher/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
supabase==2.16.0
pydantic==2.11.7
pydantic-settings==2.10.1
pytest==8.4.1
sqlalchemy==2.0.43
httpx==0.28.1
python-jose[cryptography]==3.5.0
email-validator==2.2.0
anthropic==0.52.0
python-multipart==0.0.20
